+++ jgclark.DashboardReact/changelog.md
@@ -11,6 +11,8 @@ Misc TODO:
 
 ## [2.0.0.a23]
 - fixed bug in overdues not respecting ignoreFolders on refresh
+- fixed top-alignment spacing for calendar picker
+- changed project dialog to have one month calendar picker so that it fits in the smaller dialog
 
 ## [2.0.0.a22] @jgc
 - completed layout in Project dialog (apart from CalendarPicker)